+++ app/src/main/AndroidManifest.xml
@@ -27,7 +27,6 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity android:name=".LeaderboardActivity"></activity>
     </application>
 
 </manifest>
