+++ app/src/main/res/layout/fragment_leaderboard.xml
@@ -4,10 +4,11 @@
     android:layout_height="match_parent"
     tools:context="com.nitmz.morphosis.scoobydoo.LeaderboardFragment">
 
-    <!-- TODO: Update blank fragment layout -->
-    <TextView
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:text="@string/hello_blank_fragment" />
+    <ListView
+        android:layout_width="368dp"
+        android:layout_height="495dp"
+        tools:layout_editor_absoluteX="8dp"
+        tools:layout_editor_absoluteY="8dp"
+        android:id="@+id/leaderboard_list"/>
 
 </FrameLayout>
